--- IoTBayPrototype/src/java/uts/isd/group30/model/DBConnector.java
@@ -1,46 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package uts.isd.group30.model;
-
- 
-
-import java.sql.Connection;
-
-import java.sql.DriverManager;
-
-import java.sql.SQLException;
-
- 
-
-public class DBConnector extends DB{
-
- 
-
-public DBConnector() throws ClassNotFoundException, SQLException {
-
-Class.forName(driver);
-
-conn = DriverManager.getConnection(URL, dbuser, dbpass);
-
-}
-
- 
-
-public Connection openConnection(){
-
-return this.conn;
-
-}
-
- 
-
-public void closeConnection() throws SQLException {
-
-this.conn.close();
-
-}
-
-}
